fix(FormattedNumber): rename component NumberView -> FormattedNumber
