+++ components/formatted-number/src/FootnoteView.module.css
