+++ components/formatted-number/src/FormattedNumberView.module.css
