+++ components/formatted-number/src/__stories__/FormattedNumberView.stories.module.css
