+++ components/formatted-number/src/__tests__/getMaxFractionalPartOfNumbers.tests.ts
@@ -1,6 +1,6 @@
 import { getMaxFractionalPartOfNumbers } from '../index';
 
-describe('components/NumberView/getMaxFractionalPartOfNumbers', () => {
+describe('components/FormattedNumber/getMaxFractionalPartOfNumbers', () => {
     test('max fractional part of numbers should be correct', () => {
         const numbers = [123.123, 789.2, 321];
