+++ components/formatted-number/src/footnoteEntities.ts
