+++ components/formatted-number/src/getMaxFractionalPartOfNumbers.ts
