+++ components/formatted-number/src/index.ts
@@ -1,2 +1,2 @@
-export { default } from './NumberView';
+export { default } from './FormattedNumberView';
 export { default as getMaxFractionalPartOfNumbers } from './getMaxFractionalPartOfNumbers';
