+++ components/formatted-number/tsconfig.json
